Refactored till errors
