+++ julia/constants.jl
@@ -1,0 +1,9 @@
+
+const maxdegree = 2
+const actualMaxsize = maxsize + maxdegree
+const len = size(X)[1]
+
+const nuna = size(unaops)[1]
+const nbin = size(binops)[1]
+const nops = nuna + nbin
+const nvar = size(X)[2];
