+++ julia/sr.jl
@@ -2,49 +2,10 @@ import Optim
 import Printf: @printf
 import Random: shuffle!, randperm
 
-const maxdegree = 2
-const actualMaxsize = maxsize + maxdegree
 
+include("constants.jl")
 
-# Sum of square error between two arrays
-function SSE(x::Array{Float32}, y::Array{Float32})::Float32
-    diff = (x - y)
-    return sum(diff .* diff)
-end
-function SSE(x::Nothing, y::Array{Float32})::Float32
-    return 1f9
-end
-
-# Sum of square error between two arrays, with weights
-function SSE(x::Array{Float32}, y::Array{Float32}, w::Array{Float32})::Float32
-    diff = (x - y)
-    return sum(diff .* diff .* w)
-end
-function SSE(x::Nothing, y::Array{Float32}, w::Array{Float32})::Float32
-    return Nothing
-end
-
-# Mean of square error between two arrays
-function MSE(x::Nothing, y::Array{Float32})::Float32
-    return 1f9
-end
-
-# Mean of square error between two arrays
-function MSE(x::Array{Float32}, y::Array{Float32})::Float32
-    return SSE(x, y)/size(x)[1]
-end
-
-# Mean of square error between two arrays
-function MSE(x::Nothing, y::Array{Float32}, w::Array{Float32})::Float32
-    return 1f9
-end
-
-# Mean of square error between two arrays
-function MSE(x::Array{Float32}, y::Array{Float32}, w::Array{Float32})::Float32
-    return SSE(x, y, w)/sum(w)
-end
-
-const len = size(X)[1]
+include("errors.jl")
 
 if weighted
     const avgy = sum(y .* weights)/sum(weights)
@@ -59,10 +20,7 @@ function id(x::Float32)::Float32
     x
 end
 
-const nuna = size(unaops)[1]
-const nbin = size(binops)[1]
-const nops = nuna + nbin
-const nvar = size(X)[2];
+
 
 function debug(verbosity, string...)
     verbosity > 0 ? println(string...) : nothing
